third_party/blink/web_tests/external/wpt/css/css-contain/container-queries/container-units-ineligible-container-expected.txt

This is a testharness.js-based test.
PASS /* basic */
FAIL display: table assert_equals: width expected "30px" but got "20px"
FAIL display: table-cell assert_equals: width expected "30px" but got "20px"
FAIL display: inline assert_equals: width expected "30px" but got "20px"
FAIL display: contents assert_equals: width expected "30px" but got "0px"
FAIL display: none assert_equals: width expected "30px" but got "80px"
PASS container-type: normal
PASS container-type: inline-size
PASS container-type: inline-size; writing-mode: vertical-lr
Harness: the test ran to completion.
